refactor: move resource logic to provider/publiccloud
