+++ internal/provider/provider.go
@@ -13,7 +13,6 @@ import (
 	"github.com/hashicorp/terraform-plugin-log/tflog"
 	"github.com/leaseweb/terraform-provider-leaseweb/internal/provider/client"
 	"github.com/leaseweb/terraform-provider-leaseweb/internal/provider/publiccloud"
-	"github.com/leaseweb/terraform-provider-leaseweb/internal/provider/resources/public_cloud/instance"
 )
 
 var (
@@ -168,7 +167,7 @@ func (p *leasewebProvider) DataSources(_ context.Context) []func() datasource.Da
 
 func (p *leasewebProvider) Resources(_ context.Context) []func() resource.Resource {
 	return []func() resource.Resource{
-		instance.NewInstanceResource,
+		publiccloud.NewInstanceResource,
 		NewDedicatedServerResource,
 		NewDedicatedServerCredentialResource,
 		NewDataTrafficNotificationSettingResource,
