+++ internal/provider/publiccloud/models/resource/contract_test.go
@@ -1,4 +1,4 @@
-package model
+package resource
 
 import (
 	"context"
