+++ internal/provider/publiccloud/models/resource/doc.go
@@ -1,0 +1,2 @@
+// Package resource implements models for publiccloud resources.
+package resource
