+++ internal/provider/publiccloud/resource_instance_test.go
@@ -1,4 +1,4 @@
-package instance
+package publiccloud
 
 import (
 	"context"
