+++ internal/provider/publiccloud/validator/contract_term_validator.go
@@ -7,8 +7,8 @@ import (
 	"github.com/hashicorp/terraform-plugin-framework-validators/helpers/validatordiag"
 	"github.com/hashicorp/terraform-plugin-framework/schema/validator"
 	"github.com/hashicorp/terraform-plugin-framework/types/basetypes"
+	"github.com/leaseweb/terraform-provider-leaseweb/internal/provider/publiccloud/models/resource"
 	"github.com/leaseweb/terraform-provider-leaseweb/internal/provider/publiccloud/service/public_cloud"
-	"github.com/leaseweb/terraform-provider-leaseweb/internal/provider/resources/public_cloud/model"
 )
 
 var _ validator.Object = contractTermValidator{}
@@ -32,7 +32,7 @@ func (v contractTermValidator) ValidateObject(
 ) {
 	service := public_cloud.Service{}
 
-	contract := model.Contract{}
+	contract := resource.Contract{}
 	request.ConfigValue.As(ctx, &contract, basetypes.ObjectAsOptions{})
 	err := service.ValidateContractTerm(
 		contract.Term.ValueInt64(),
