--- internal/provider/resources/public_cloud/instance/create.go
@@ -1,46 +1,0 @@
-package instance
-
-import (
-	"context"
-
-	"github.com/hashicorp/terraform-plugin-framework/resource"
-	"github.com/hashicorp/terraform-plugin-log/tflog"
-	"github.com/leaseweb/terraform-provider-leaseweb/internal/provider/resources/public_cloud/model"
-	"github.com/leaseweb/terraform-provider-leaseweb/internal/provider/shared/logging"
-)
-
-func (i *instanceResource) Create(
-	ctx context.Context,
-	req resource.CreateRequest,
-	resp *resource.CreateResponse,
-) {
-	var plan model.Instance
-
-	diags := req.Plan.Get(ctx, &plan)
-	resp.Diagnostics.Append(diags...)
-	if resp.Diagnostics.HasError() {
-		return
-	}
-
-	tflog.Info(ctx, "Creating public cloud instance")
-	instance, err := i.client.PublicCloudService.LaunchInstance(plan, ctx)
-	if err != nil {
-		resp.Diagnostics.AddError("Error creating Instance", err.Error())
-
-		logging.ServiceError(
-			ctx,
-			err.ErrorResponse,
-			&resp.Diagnostics,
-			"Error creating public cloud instance",
-			err.Error(),
-		)
-
-		return
-	}
-
-	diags = resp.State.Set(ctx, instance)
-	resp.Diagnostics.Append(diags...)
-	if resp.Diagnostics.HasError() {
-		return
-	}
-}
