--- internal/provider/resources/public_cloud/instance/doc.go
@@ -1,2 +1,0 @@
-// Package instance implements functionality for public_cloud resources.
-package instance
