--- internal/provider/resources/public_cloud/instance/schema.go
@@ -1,126 +1,0 @@
-package instance
-
-import (
-	"context"
-
-	"github.com/hashicorp/terraform-plugin-framework-validators/int64validator"
-	"github.com/hashicorp/terraform-plugin-framework-validators/stringvalidator"
-	"github.com/hashicorp/terraform-plugin-framework/path"
-	"github.com/hashicorp/terraform-plugin-framework/resource"
-	"github.com/hashicorp/terraform-plugin-framework/resource/schema"
-	"github.com/hashicorp/terraform-plugin-framework/resource/schema/planmodifier"
-	"github.com/hashicorp/terraform-plugin-framework/resource/schema/stringplanmodifier"
-	"github.com/hashicorp/terraform-plugin-framework/schema/validator"
-	"github.com/leaseweb/terraform-provider-leaseweb/internal/provider/publiccloud/service/public_cloud"
-	sharedSchemas "github.com/leaseweb/terraform-provider-leaseweb/internal/provider/shared_schemas/public_cloud"
-)
-
-func (i *instanceResource) Schema(
-	_ context.Context,
-	_ resource.SchemaRequest,
-	resp *resource.SchemaResponse,
-) {
-	service := public_cloud.Service{}
-	warningError := "**WARNING!** Changing this value once running will cause this instance to be destroyed and a new one to be created."
-
-	resp.Schema = schema.Schema{
-		Attributes: map[string]schema.Attribute{
-			"id": schema.StringAttribute{
-				Computed:    true,
-				Description: "The instance unique identifier",
-				PlanModifiers: []planmodifier.String{
-					stringplanmodifier.UseStateForUnknown(),
-				},
-			},
-			"region": schema.StringAttribute{
-				Required:    true,
-				Description: "Our current regions can be found in the [developer documentation](https://developer.leaseweb.com/api-docs/publiccloud_v1.html#tag/Instances/operation/launchInstance)" + warningError,
-				PlanModifiers: []planmodifier.String{
-					stringplanmodifier.RequiresReplace(),
-				},
-			},
-			"reference": schema.StringAttribute{
-				Optional:    true,
-				Computed:    true,
-				Description: "The identifying name set to the instance",
-			},
-			"image": schema.SingleNestedAttribute{
-				Required: true,
-				Attributes: map[string]schema.Attribute{
-					"id": schema.StringAttribute{
-						Required:    true,
-						Description: "Can be either an Operating System or a UUID in case of a Custom Image ID." + warningError,
-						PlanModifiers: []planmodifier.String{
-							stringplanmodifier.RequiresReplace(),
-						},
-					},
-				},
-			},
-			"state": schema.StringAttribute{
-				Computed:    true,
-				Description: "The instance's current state",
-			},
-			"type": schema.StringAttribute{
-				Required: true,
-				Validators: []validator.String{
-					stringvalidator.AlsoRequires(
-						path.Expressions{path.MatchRoot("region")}...,
-					),
-				},
-			},
-			// TODO Enable SSH key support
-			/**
-			  "ssh_key": schema.StringAttribute{
-			  	Optional:      true,
-			  	Sensitive:     true,
-			  	PlanModifiers: []planmodifier.String{stringplanmodifier.UseStateForUnknown()},
-			  	Description:   "Public SSH key to be installed into the instance. Must be used only on Linux/FreeBSD instances",
-			  	Validators: []validator.String{
-			  		stringvalidator.RegexMatches(
-			  			regexp.MustCompile(service.GetSshKeyRegularExpression()),
-			  			"Invalid ssh key",
-			  		),
-			  	},
-			  },
-			*/
-			"root_disk_size": schema.Int64Attribute{
-				Computed:    true,
-				Optional:    true,
-				Description: "The root disk's size in GB. Must be at least 5 GB for Linux and FreeBSD instances and 50 GB for Windows instances. The maximum size is 1000 GB",
-				Validators: []validator.Int64{
-					int64validator.Between(
-						service.GetMinimumRootDiskSize(),
-						service.GetMaximumRootDiskSize(),
-					),
-				},
-			},
-			"root_disk_storage_type": schema.StringAttribute{
-				Required:    true,
-				Description: "The root disk's storage type. Can be *LOCAL* or *CENTRAL*. " + warningError,
-				Validators: []validator.String{
-					stringvalidator.OneOf(service.GetRootDiskStorageTypes()...),
-				},
-				PlanModifiers: []planmodifier.String{
-					stringplanmodifier.RequiresReplace(),
-				},
-			},
-			"ips": schema.ListNestedAttribute{
-				Computed: true,
-				NestedObject: schema.NestedAttributeObject{
-					Attributes: map[string]schema.Attribute{
-						"ip": schema.StringAttribute{Computed: true},
-					},
-				},
-			},
-			"contract": sharedSchemas.Contract(true, &service),
-			"market_app_id": schema.StringAttribute{
-				Computed:    true,
-				Optional:    true,
-				Description: "Market App ID that must be installed into the instance." + warningError,
-				PlanModifiers: []planmodifier.String{
-					stringplanmodifier.RequiresReplaceIfConfigured(),
-				},
-			},
-		},
-	}
-}
