--- internal/provider/resources/public_cloud/model/doc.go
@@ -1,2 +1,0 @@
-// Package model implements models for public_cloud resources.
-package model
